+++ tests/test_sloan_log.py
@@ -15,34 +15,37 @@ class TestSloanLog(unittest.TestCase):
         self.assertTrue(Path(ap_data_dir).exists())
         self.assertTrue(Path(b_data_dir).exists())
 
-    def test_known_data(self):
-        """Runs on an old dataset that I know used to run successfully"""
+    def setUp(self) -> None:
 
         class Dummy(object):
             pass
 
-        args = Dummy()
-        args.sjd = 59011
-        args.print = True
-        args.summary = True
-        args.data = True
-        args.boss = True
-        args.apogee = True
-        args.p_boss = True
-        args.p_apogee = True
-        args.log_support = True
-        args.morning = False
-        args.verbose = True
-        args.noprogress = False
-        args.mirrors = True
-        args.telstatus = True
-        args.legacy_aptest = False
+        self.args = Dummy()
 
-        ap_data_dir = sloan_log.ap_dir / '{}'.format(args.sjd)
-        b_data_dir = sloan_log.b_dir / '{}'.format(args.sjd)
+    def test_known_data(self):
+        """Runs on an old dataset that I know used to run successfully"""
+
+        self.args.sjd = 59011
+        self.args.print = True
+        self.args.summary = True
+        self.args.data = True
+        self.args.boss = True
+        self.args.apogee = True
+        self.args.p_boss = True
+        self.args.p_apogee = True
+        self.args.log_support = True
+        self.args.morning = False
+        self.args.verbose = True
+        self.args.noprogress = False
+        self.args.mirrors = True
+        self.args.telstatus = True
+        self.args.legacy_aptest = False
+
+        ap_data_dir = sloan_log.ap_dir / '{}'.format(self.args.sjd)
+        b_data_dir = sloan_log.b_dir / '{}'.format(self.args.sjd)
         ap_images = list(Path(ap_data_dir).glob('apR-a*.apz'))
         b_images = list(Path(b_data_dir).glob('sdR-r1*fit.gz'))
-        log = sloan_log.Logging(ap_images, b_images, args)
+        log = sloan_log.Logging(ap_images, b_images, self.args)
         log.parse_images()
         log.sort()
         log.count_dithers()
@@ -57,30 +60,26 @@ class TestSloanLog(unittest.TestCase):
         only runs BOSS data summary
         """
 
-        class Dummy(object):
-            pass
+        self.args.sjd = 59011
+        self.args.print = False
+        self.args.summary = False
+        self.args.data = False
+        self.args.boss = True
+        self.args.apogee = False
+        self.args.p_boss = True
+        self.args.p_apogee = False
+        self.args.log_support = False
+        self.args.morning = False
+        self.args.verbose = True
+        self.args.mirrors = False
+        self.args.telstatus = False
+        self.args.legacy_aptest = False
 
-        args = Dummy()
-        args.sjd = 59011
-        args.print = False
-        args.summary = False
-        args.data = False
-        args.boss = True
-        args.apogee = False
-        args.p_boss = True
-        args.p_apogee = False
-        args.log_support = False
-        args.morning = False
-        args.verbose = True
-        args.mirrors = False
-        args.telstatus = False
-        args.legacy_aptest = False
-
-        ap_data_dir = sloan_log.ap_dir / '{}'.format(args.sjd)
-        b_data_dir = sloan_log.b_dir / '{}'.format(args.sjd)
+        ap_data_dir = sloan_log.ap_dir / '{}'.format(self.args.sjd)
+        b_data_dir = sloan_log.b_dir / '{}'.format(self.args.sjd)
         ap_images = list(Path(ap_data_dir).glob('apR-a*.apz'))
         b_images = list(Path(b_data_dir).glob('sdR-r1*fit.gz'))
-        log = sloan_log.Logging(ap_images, b_images, args)
+        log = sloan_log.Logging(ap_images, b_images, self.args)
         log.parse_images()
         log.sort()
         log.count_dithers()
@@ -91,30 +90,26 @@ class TestSloanLog(unittest.TestCase):
         only checks APOGEE and prints its summary
         """
 
-        class Dummy(object):
-            pass
+        self.args.sjd = 59011
+        self.args.print = False
+        self.args.summary = False
+        self.args.data = False
+        self.args.boss = False
+        self.args.apogee = True
+        self.args.p_boss = False
+        self.args.p_apogee = True
+        self.args.log_support = False
+        self.args.morning = False
+        self.args.verbose = True
+        self.args.morrors = False
+        self.args.telstatus = False
+        self.args.legacy_aptest = False
 
-        args = Dummy()
-        args.sjd = 59011
-        args.print = False
-        args.summary = False
-        args.data = False
-        args.boss = False
-        args.apogee = True
-        args.p_boss = False
-        args.p_apogee = True
-        args.log_support = False
-        args.morning = False
-        args.verbose = True
-        args.morrors = False
-        args.telstatus = False
-        args.legacy_aptest = False
-
-        ap_data_dir = sloan_log.ap_dir / '{}'.format(args.sjd)
-        b_data_dir = sloan_log.b_dir / '{}'.format(args.sjd)
+        ap_data_dir = sloan_log.ap_dir / '{}'.format(self.args.sjd)
+        b_data_dir = sloan_log.b_dir / '{}'.format(self.args.sjd)
         ap_images = list(Path(ap_data_dir).glob('apR-a*.apz'))
         b_images = list(Path(b_data_dir).glob('sdR-r1*fit.gz'))
-        log = sloan_log.Logging(ap_images, b_images, args)
+        log = sloan_log.Logging(ap_images, b_images, self.args)
         log.parse_images()
         log.sort()
         log.count_dithers()
@@ -123,30 +118,26 @@ class TestSloanLog(unittest.TestCase):
     def test_log_support(self):
         """Runs on an old dataset that I know used to run successfully"""
 
-        class Dummy(object):
-            pass
+        self.args.sjd = 59011
+        self.args.print = False
+        self.args.summary = False
+        self.args.data = False
+        self.args.boss = False
+        self.args.apogee = False
+        self.args.p_boss = False
+        self.args.p_apogee = False
+        self.args.log_support = True
+        self.args.morning = False
+        self.args.verbose = True
+        self.args.mirrors = False
+        self.args.telstatus = False
+        self.args.legacy_aptest = False
 
-        args = Dummy()
-        args.sjd = 59011
-        args.print = False
-        args.summary = False
-        args.data = False
-        args.boss = False
-        args.apogee = False
-        args.p_boss = False
-        args.p_apogee = False
-        args.log_support = True
-        args.morning = False
-        args.verbose = True
-        args.mirrors = False
-        args.telstatus = False
-        args.legacy_aptest = False
-
-        ap_data_dir = sloan_log.ap_dir / '{}'.format(args.sjd)
-        b_data_dir = sloan_log.b_dir / '{}'.format(args.sjd)
+        ap_data_dir = sloan_log.ap_dir / '{}'.format(self.args.sjd)
+        b_data_dir = sloan_log.b_dir / '{}'.format(self.args.sjd)
         ap_images = list(Path(ap_data_dir).glob('apR-a*.apz'))
         b_images = list(Path(b_data_dir).glob('sdR-r1*fit.gz'))
-        log = sloan_log.Logging(ap_images, b_images, args)
+        log = sloan_log.Logging(ap_images, b_images, self.args)
         log.parse_images()
         log.sort()
         log.count_dithers()
@@ -155,30 +146,26 @@ class TestSloanLog(unittest.TestCase):
     def test_log_support_today(self):
         """Runs on an old dataset that I know used to run successfully"""
 
-        class Dummy(object):
-            pass
+        self.args.sjd = sjd.sjd()
+        self.args.print = False
+        self.args.summary = False
+        self.args.data = False
+        self.args.boss = False
+        self.args.apogee = False
+        self.args.p_boss = False
+        self.args.p_apogee = False
+        self.args.log_support = True
+        self.args.morning = False
+        self.args.verbose = True
+        self.args.mirrors = False
+        self.args.telstatus = False
+        self.args.legacy_aptest = False
 
-        args = Dummy()
-        args.sjd = sjd.sjd()
-        args.print = False
-        args.summary = False
-        args.data = False
-        args.boss = False
-        args.apogee = False
-        args.p_boss = False
-        args.p_apogee = False
-        args.log_support = True
-        args.morning = False
-        args.verbose = True
-        args.mirrors = False
-        args.telstatus = False
-        args.legacy_aptest = False
-
-        ap_data_dir = sloan_log.ap_dir / '{}'.format(args.sjd)
-        b_data_dir = sloan_log.b_dir / '{}'.format(args.sjd)
+        ap_data_dir = sloan_log.ap_dir / '{}'.format(self.args.sjd)
+        b_data_dir = sloan_log.b_dir / '{}'.format(self.args.sjd)
         ap_images = list(Path(ap_data_dir).glob('apR-a*.apz'))
         b_images = list(Path(b_data_dir).glob('sdR-r1*fit.gz'))
-        log = sloan_log.Logging(ap_images, b_images, args)
+        log = sloan_log.Logging(ap_images, b_images, self.args)
         log.parse_images()
         log.sort()
         log.count_dithers()
@@ -187,30 +174,26 @@ class TestSloanLog(unittest.TestCase):
     def test_mirror_numbers(self):
         """Runs on an old dataset that I know used to run successfully"""
 
-        class Dummy(object):
-            pass
+        self.args.sjd = sjd.sjd()
+        self.args.print = False
+        self.args.summary = False
+        self.args.data = False
+        self.args.boss = False
+        self.args.apogee = False
+        self.args.p_boss = False
+        self.args.p_apogee = False
+        self.args.log_support = False
+        self.args.morning = False
+        self.args.verbose = True
+        self.args.mirrors = True
+        self.args.telstatus = False
+        self.args.legacy_aptest = False
 
-        args = Dummy()
-        args.sjd = sjd.sjd()
-        args.print = False
-        args.summary = False
-        args.data = False
-        args.boss = False
-        args.apogee = False
-        args.p_boss = False
-        args.p_apogee = False
-        args.log_support = False
-        args.morning = False
-        args.verbose = True
-        args.mirrors = True
-        args.telstatus = False
-        args.legacy_aptest = False
-
-        ap_data_dir = sloan_log.ap_dir / '{}'.format(args.sjd)
-        b_data_dir = sloan_log.b_dir / '{}'.format(args.sjd)
+        ap_data_dir = sloan_log.ap_dir / '{}'.format(self.args.sjd)
+        b_data_dir = sloan_log.b_dir / '{}'.format(self.args.sjd)
         ap_images = list(Path(ap_data_dir).glob('apR-a*.apz'))
         b_images = list(Path(b_data_dir).glob('sdR-r1*fit.gz'))
-        log = sloan_log.Logging(ap_images, b_images, args)
+        log = sloan_log.Logging(ap_images, b_images, self.args)
         log.parse_images()
         log.sort()
         log.count_dithers()
@@ -221,35 +204,63 @@ class TestSloanLog(unittest.TestCase):
 def test_tel_status(self):
     """Test telstatus integration"""
 
-    class Dummy(object):
-        pass
+    self.args.sjd = sjd.sjd()
+    self.args.print = False
+    self.args.summary = False
+    self.args.data = False
+    self.args.boss = False
+    self.args.apogee = False
+    self.args.p_boss = False
+    self.args.p_apogee = False
+    self.args.log_support = False
+    self.args.morning = False
+    self.args.verbose = True
+    self.args.mirrors = False
+    self.args.telstatus = True
+    self.args.legacy_aptest = False
 
-    args = Dummy()
-    args.sjd = sjd.sjd()
-    args.print = False
-    args.summary = False
-    args.data = False
-    args.boss = False
-    args.apogee = False
-    args.p_boss = False
-    args.p_apogee = False
-    args.log_support = False
-    args.morning = False
-    args.verbose = True
-    args.mirrors = False
-    args.telstatus = True
-    args.legacy_aptest = False
-
-    ap_data_dir = sloan_log.ap_dir / '{}'.format(args.sjd)
-    b_data_dir = sloan_log.b_dir / '{}'.format(args.sjd)
+    ap_data_dir = sloan_log.ap_dir / '{}'.format(self.args.sjd)
+    b_data_dir = sloan_log.b_dir / '{}'.format(self.args.sjd)
     ap_images = list(Path(ap_data_dir).glob('apR-a*.apz'))
     b_images = list(Path(b_data_dir).glob('sdR-r1*fit.gz'))
-    log = sloan_log.Logging(ap_images, b_images, args)
+    log = sloan_log.Logging(ap_images, b_images, self.args)
     log.parse_images()
     log.sort()
     log.count_dithers()
     log.log_support()
-    log.mirror_numbers()
+
+
+def test_legacy_aptest(self):
+    """Test utr_cdr-based aptest that matches the old aptest script"""
+
+    class Dummy(object):
+        pass
+
+    args = Dummy()
+    self.args.sjd = sjd.sjd()
+    self.args.print = False
+    self.args.summary = False
+    self.args.data = True
+    self.args.boss = False
+    self.args.apogee = False
+    self.args.p_boss = False
+    self.args.p_apogee = False
+    self.args.log_support = False
+    self.args.morning = False
+    self.args.verbose = True
+    self.args.mirrors = False
+    self.args.telstatus = False
+    self.args.legacy_aptest = True
+
+    ap_data_dir = sloan_log.ap_dir / '{}'.format(self.args.sjd)
+    b_data_dir = sloan_log.b_dir / '{}'.format(self.args.sjd)
+    ap_images = list(Path(ap_data_dir).glob('apR-a*.apz'))
+    b_images = list(Path(b_data_dir).glob('sdR-r1*fit.gz'))
+    log = sloan_log.Logging(ap_images, b_images, self.args)
+    log.parse_images()
+    log.sort()
+    log.count_dithers()
+    log.p_data()
 
 
 if __name__ == '__main__':